requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
geopy==2.4.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
openai==1.3.5
PyPDF2==3.0.1
